Changed prime function to return the smallest divisor if the number is not prime
And I'm wondering if the file .idea/workspace.xml should be tracked in git..
It is simply dumping the diff with unnecessary changes.

Maybe owing to company policy, this will contribute to unnecessariness of the project. So we shall retain the file for tracking.

Unnecessary President signing off. Thank you!